update commands work. show action works.

+++ japkit-roo-quickstart-example/src/main/java/de/stefanocke/japkit/roo/japkit/domain/IdAndVersion.java
@@ -31,6 +31,6 @@ public class IdAndVersion {
 	@Field
 	@Version
 	@Column(name = "version")
-	private Integer version;
+	private Long version;
 
 }
